+++ public/index.html
@@ -36,14 +36,13 @@
     <!-- Open Graph Meta Tags (Facebook, LinkedIn, etc.) -->
   <meta property="og:title" content="Home | About | Rhythm Messengers">
   <meta property="og:description" content="Welcome to The Rhythm Messengers, the leading entertainment for events.">
-  <meta property="og:image" content="https://https://rhythmmessengers.com/media/RMELogo300-2-no-background.png">
-  <meta property="og:url" content="https://https://rhythmmessengers.com//">
+  <meta property="og:image" content="https://rhythmmessengers.com/media/RMELogo300-2-no-background.png">
+  <meta property="og:url" content="https://rhythmmessengers.com">
   <meta property="og:type" content="website">
   <meta property="og:site_name" content="The Rhythm Messengers">
   <meta property="og:locale" content="en_US">
   
   <!-- Apple Mobile Web App Meta Tags -->
-  <meta name="apple-mobile-web-app-capable" content="yes"> <!-- Enable standalone mode on iOS -->
   <meta name="apple-mobile-web-app-title" content="The Rhythm Messengers"> <!-- App name for home screen -->
   <meta name="apple-mobile-web-app-status-bar-style" content="black-translucent"> <!-- Status bar style -->
   
@@ -51,10 +50,10 @@
   <meta name="theme-color" content="#000000"> <!-- Set the theme color for Android taskbar (Chrome on Android) -->
 
   <!-- Icon for Apple Touch (Icon when saved to Home Screen) -->
-  <link rel="apple-touch-icon" href="https://https://rhythmmessengers.com/media/RMELogo300-2-no-background.png">
+  <link rel="apple-touch-icon" href="https://rhythmmessengers.com/media/RMELogo300-2-no-background.png">
 
   <!-- Icon for Android/General (Icon when saved to Home Screen) -->
-  <link rel="icon" href="https://https://rhythmmessengers.com/media/RMELogo300-2-no-background.png" type="image/x-icon">
+  <link rel="icon" href="https://rhythmmessengers.com/media/RMELogo300-2-no-background.png" type="image/x-icon">
 
     <title>Rhythm Messengers</title>
   </head>
